+++ application/controllers/Home.php
@@ -1,7 +1,27 @@
-<?php defined('BASEPATH') OR exit('No direct script access allowed');
+<?php
+/**
+ * A3M (Account Authentication & Authorization) is a CodeIgniter 3.x package.
+ * It gives you the CRUD to get working right away without too much fuss and tinkering!
+ * Designed for building webapps from scratch without all that tiresome login / logout / admin stuff thats always required.
+ *
+ * @link https://github.com/donjakobo/A3M GitHub repository
+ */
+if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
-class Home extends CI_Controller {
-
+/**
+ * Static pages
+ *
+ * This controller is for the home page and any other top level static pages.
+ * 
+ * @package A3M
+ * @subpackage Controllers
+ */
+class Home extends CI_Controller
+{
+	
+	/**
+	 * Constructor
+	 */
 	function __construct()
 	{
 		parent::__construct();
@@ -11,15 +31,15 @@ class Home extends CI_Controller {
 		$this->load->library(array('account/authentication', 'account/authorization'));
 		$this->load->model('account/Account_model');
 	}
-
+	
+	/**
+	 * Home page
+	 *
+	 * This is the controller for the home page
+	 */
 	function index()
 	{
-		maintain_ssl();
-		
-		if ($this->authentication->is_signed_in())
-		{
-			$data['account'] = $this->Account_model->get_by_id($this->session->userdata('account_id'));
-		}
+		$data = $this->authentication->initialize(FALSE);
 		
 		$data['content'] = $this->load->view('home', isset($data) ? $data : NULL, TRUE);
 		$this->load->view('template', $data);