+++ application/controllers/account/Connect.php
@@ -1,6 +1,16 @@
-<?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
-/*
- * Connect Controller
+<?php
+/**
+ * A3M (Account Authentication & Authorization) is a CodeIgniter 3.x package.
+ * It gives you the CRUD to get working right away without too much fuss and tinkering!
+ * Designed for building webapps from scratch without all that tiresome login / logout / admin stuff thats always required.
+ *
+ * @link https://github.com/donjakobo/A3M GitHub repository
+ */
+if ( ! defined('BASEPATH')) exit('No direct script access allowed');
+/**
+ * Connect to social providers
+ * @package A3M
+ * @subpackage Controllers
  */
 class Connect extends CI_Controller
 {
@@ -19,6 +29,11 @@ class Connect extends CI_Controller
         $this->load->language(array('general', 'account/connect_third_party'));
     }
     
+    /**
+     * The initiation of connecting to a third party provider
+     * @param string $provider Name of the provider that we should connect to
+     * @param string $identifier For OpenID, the link to the account
+     */
     function Index($provider = NULL, $identifier = NULL)
     {
         // Enable SSL?