+++ application/controllers/account/Sign_in.php
@@ -1,8 +1,20 @@
-<?php defined('BASEPATH') OR exit('No direct script access allowed');
-/*
- * Sign_in Controller
+<?php
+/**
+ * A3M (Account Authentication & Authorization) is a CodeIgniter 3.x package.
+ * It gives you the CRUD to get working right away without too much fuss and tinkering!
+ * Designed for building webapps from scratch without all that tiresome login / logout / admin stuff thats always required.
+ *
+ * @link https://github.com/donjakobo/A3M GitHub repository
  */
-class Sign_in extends CI_Controller {
+defined('BASEPATH') OR exit('No direct script access allowed');
+/**
+ * Sign in
+ *
+ * @package A3M
+ * @subpackage Controllers
+ */
+class Sign_in extends CI_Controller
+{
 
 	/**
 	 * Constructor