+++ application/models/account/Account_details_model.php
@@ -1,6 +1,23 @@
-<?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
+<?php
+/**
+ * A3M (Account Authentication & Authorization) is a CodeIgniter 3.x package.
+ * It gives you the CRUD to get working right away without too much fuss and tinkering!
+ * Designed for building webapps from scratch without all that tiresome login / logout / admin stuff thats always required.
+ *
+ * @link https://github.com/donjakobo/A3M GitHub repository
+ */
+if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
-class Account_details_model extends CI_Model {
+/**
+ * Account_details_model
+ *
+ * Model to retrieve and update more detailed information about user.
+ *
+ * @package A3M
+ * @subpackage Models
+ */
+class Account_details_model extends CI_Model
+{
 
 	/**
 	 * Get details for all account
@@ -10,7 +27,7 @@ class Account_details_model extends CI_Model {
 	 */
 	function get()
 	{
-		return $this->db->get('a3m_account_details')->result();
+		return $this->db->get($this->db->dbprefix . 'a3m_account_details')->result();
 	}
 
 	/**
@@ -22,7 +39,7 @@ class Account_details_model extends CI_Model {
 	 */
 	function get_by_account_id($account_id)
 	{
-		return $this->db->get_where('a3m_account_details', array('account_id' => $account_id))->row();
+		return $this->db->get_where($this->db->dbprefix . 'a3m_account_details', array('account_id' => $account_id))->row();
 	}
 
 	// --------------------------------------------------------------------
@@ -65,14 +82,6 @@ class Account_details_model extends CI_Model {
 			$country = $this->ref_country_model->get($attributes['country']);
 			$country ? $attributes['country'] = $country->alpha2 : NULL;
 		}
-		
-		// @todo This needs to be replaced by something more sensible as the default database is too big with this.
-		// At this point, if country is still not determined, use ip address to determine country
-		if ( ! isset($attributes['country']))
-		{
-			
-		}
-		
 		// Check that it's a recognized language (see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes)
 		if (isset($attributes['language']))
 		{
@@ -110,17 +119,16 @@ class Account_details_model extends CI_Model {
 		if ($this->get_by_account_id($account_id))
 		{
 			$this->db->where('account_id', $account_id);
-			$this->db->update('a3m_account_details', $attributes);
+			$this->db->update($this->db->dbprefix . 'a3m_account_details', $attributes);
 		}
 		// Insert
 		else
 		{
 			$attributes['account_id'] = $account_id;
-			$this->db->insert('a3m_account_details', $attributes);
+			$this->db->insert($this->db->dbprefix . 'a3m_account_details', $attributes);
 		}
 	}
 
 }
-
 /* End of file Account_details_model.php */
 /* Location: ./application/models/account/Account_details_model.php */