+++ application/models/account/Ref_zoneinfo_model.php
@@ -1,6 +1,24 @@
-<?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
+<?php
+/**
+ * A3M (Account Authentication & Authorization) is a CodeIgniter 3.x package.
+ * It gives you the CRUD to get working right away without too much fuss and tinkering!
+ * Designed for building webapps from scratch without all that tiresome login / logout / admin stuff thats always required.
+ *
+ * @link https://github.com/donjakobo/A3M GitHub repository
+ */
+if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
-class Ref_zoneinfo_model extends CI_Model {
+/**
+ * Ref_zoneinfo
+ *
+ * Model for the Ref_zoneinfo table.
+ * Refferencing timezone information.
+ *
+ * @package A3M
+ * @subpackage Models
+ */
+class Ref_zoneinfo_model extends CI_Model
+{
 
 	/**
 	 * Get by zoneinfo
@@ -12,7 +30,7 @@ class Ref_zoneinfo_model extends CI_Model {
 	function get_by_zoneinfo($zoneinfo)
 	{
 		$this->db->where('zoneinfo', $zoneinfo);
-		$query = $this->db->get('ref_zoneinfo');
+		$query = $this->db->get($this->db->dbprefix . 'ref_zoneinfo');
 		if ($query->num_rows()) return $query->row();
 	}
 
@@ -28,7 +46,7 @@ class Ref_zoneinfo_model extends CI_Model {
 	function get_by_country($country)
 	{
 		$this->db->where('country', $country);
-		$query = $this->db->get('ref_zoneinfo');
+		$query = $this->db->get($this->db->dbprefix . 'ref_zoneinfo');
 		if ($query->num_rows()) return $query->result();
 	}
 
@@ -43,11 +61,9 @@ class Ref_zoneinfo_model extends CI_Model {
 	function get_all()
 	{
 		$this->db->order_by('zoneinfo', 'asc');
-		return $this->db->get('ref_zoneinfo')->result();
+		return $this->db->get($this->db->dbprefix . 'ref_zoneinfo')->result();
 	}
 
 }
-
-
 /* End of file Ref_zoneinfo_model.php */
 /* Location: ./application/models/account/Ref_zoneinfo_model.php */